crux-search-for-matches --compute-p-values T now handles memory better. Rather than keep all psms generated by each peptide mod and sample from from them at the end, it samples some after each peptide mod is scored for the prelimiary score.  The randomly selected psms are kept in a new member variable of the match_collection.  Then the psms are sorted by preliminary score and truncated.
Requires a minimum of 40 sampled psms to estimate weibull parameters.
At DETAILED_INFO level (40) prints a warning if params could not be
estimated.

By setting the seed option in param file, the random selection is
reproducible so that scores are consistent for multiple runs with the
same inputs (for testing)

Smoke test updated.